+++ src/memory/ft_memdup.c
@@ -6,7 +6,7 @@
 /*   By: asoursou <asoursou@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/04/12 13:46:06 by asoursou          #+#    #+#             */
-/*   Updated: 2020/06/02 16:53:56 by asoursou         ###   ########.fr       */
+/*   Updated: 2020/08/29 19:54:33 by asoursou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,13 +14,14 @@
 #include "ft_memory.h"
 #include "ft_type.h"
 
-void	*ft_memdup(const void *src, size_t alloc_size, size_t copy_size)
+void	*ft_memdup(const void *src, size_t alloc_size, size_t copy_size,
+		bool fill)
 {
 	t_u8 *d;
 
 	if (!(d = malloc(alloc_size)))
 		return (NULL);
-	if (alloc_size > copy_size)
+	if (alloc_size > copy_size && fill)
 		ft_bzero(d + copy_size, alloc_size - copy_size);
 	else if (copy_size > alloc_size)
 		copy_size = alloc_size;