+++ src/string/ft_strdup.c
@@ -6,7 +6,7 @@
 /*   By: asoursou <asoursou@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/08/12 04:28:09 by asoursou          #+#    #+#             */
-/*   Updated: 2020/08/27 10:16:00 by asoursou         ###   ########.fr       */
+/*   Updated: 2020/08/29 19:59:10 by asoursou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,5 +18,5 @@ char	*ft_strdup(const char *s1)
 	size_t n;
 
 	n = ft_strlen(s1) + 1;
-	return (ft_memdup(s1, n, n));
+	return (ft_memdup(s1, n, n, false));
 }