+++ src/string/ft_strstr.c
@@ -6,27 +6,14 @@
 /*   By: asoursou <asoursou@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/08/12 04:56:45 by asoursou          #+#    #+#             */
-/*   Updated: 2020/08/26 16:04:47 by asoursou         ###   ########.fr       */
+/*   Updated: 2020/08/29 19:40:03 by asoursou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
+#include "ft_memory.h"
 #include "ft_string.h"
 
-char	*ft_strstr(const char *haystack, const char *needle)
+char	*ft_strstr(const char *h, const char *n)
 {
-	size_t hlen;
-	size_t nlen;
-
-	if ((nlen = ft_strlen(needle)))
-	{
-		hlen = ft_strlen(haystack);
-		while (hlen >= nlen && ft_strncmp(haystack, needle, nlen))
-		{
-			++haystack;
-			--hlen;
-		}
-		if (hlen < nlen)
-			return (NULL);
-	}
-	return ((char*)haystack);
+	return (ft_memmem(h, ft_strlen(h), n, ft_strlen(n)));
 }