+++ src/string/ft_strsub.c
@@ -6,7 +6,7 @@
 /*   By: asoursou <asoursou@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/08/15 00:54:04 by asoursou          #+#    #+#             */
-/*   Updated: 2020/08/26 15:29:04 by asoursou         ###   ########.fr       */
+/*   Updated: 2020/08/29 20:00:07 by asoursou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,5 +22,5 @@ char	*ft_strsub(const char *s, size_t start, size_t len)
 	s += start;
 	if ((n = ft_strlen(s)) > len)
 		n = len;
-	return (ft_memdup(s, n + 1, n));
+	return (ft_memdup(s, n + 1, n, true));
 }